Refactor move filter search params hook
