+++ apps/admin-ui/src/lib/application-rounds/[id]/ApplicationSectionDataLoader.tsx
@@ -8,7 +8,7 @@ import { errorToast } from "common/src/common/toast";
 import { More } from "@/component/More";
 import { useSort } from "@/hooks/useSort";
 import { ApplicationSectionsTable, SORT_KEYS } from "./ApplicationSectionsTable";
-import { useGetFilterSearchParams } from "./utils";
+import { useGetFilterSearchParams } from "@/hooks";
 import { CenterSpinner } from "common/styled";
 
 type Props = {
